+++ Base/ModLangGenBase.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using Bark.BetterCCL;
 using BepInEx.Logging;
 
@@ -8,6 +9,8 @@ public abstract class ModLangGenBase
     private bool _isInitialized;
     private ManualLogSource _log = null!;
     protected abstract string LanguageCode { get; }
+    protected abstract string NameSpace { get; }
+    public static readonly List<string> Locales = [];
 
     public int Count { get; private set; }
 
@@ -25,8 +28,9 @@ public abstract class ModLangGenBase
     protected void Add(string category, string key, string value)
     {
         if (string.IsNullOrEmpty(key)) return;
-        BetterLocale.SetDefault(LanguageCode, category, key, value);
+        BetterLocale.SetDefault(LanguageCode, category, $"{NameSpace}.{key}", value);
         Count++;
+        Locales.Add(key);
     }
 
     protected void Other(string key, string value)