+++ Plugin.cs
@@ -1,8 +1,11 @@
+using System.IO;
+using Bark.Base;
 using Bark.BetterCCL;
 using Bark.Example.Lang;
 using Bark.Tool;
 using BepInEx;
 using BepInEx.Logging;
+using CUCoreLib.Registries;
 using HarmonyLib;
 
 namespace Bark;
@@ -13,7 +16,7 @@ public class Plugin : BaseUnityPlugin
 {
     public const string Guid = "org.cncumc.bark";
     public const string Name = "Bark";
-    public const string Version = "1.0.3";
+    public const string Version = "1.1.0";
     internal new static ManualLogSource Logger = null!;
     private readonly Harmony _harmony = new(Guid);
 
@@ -21,14 +24,25 @@ public class Plugin : BaseUnityPlugin
     {
         Logger = base.Logger;
 
-        new EnLangGenerator().Initialize(Logger);
+        new ENLangGenerator().Initialize(Logger);
         new ZhCnLangGenerator().Initialize(Logger);
         new ZhTwLangGenerator().Initialize(Logger);
 
         BetterOptions.Bool("bark", "test", Setting.SettingCategory.Game, false);
         BetterLocale.Flush();
         _harmony.PatchAll();
-
+        
+        ConsoleCommandRegistry.Register(
+            "catfcabl",
+            BetterLocale.GetCommand("catfcabl"),
+            _ =>
+            {
+                var path = Paths.CachePath + "\\catfcabl.txt";
+                File.WriteAllLines(path, ModLangGenBase.Locales);
+                LogUtil.Message($"catfcabl.txt: {path}");
+            }
+        );
+        
         UpdateUtil.Check("CNCUMC/Bark", Name, Version, Logger);
     }
 }